pyproject.toml
[tool.poetry]
name = "erd-to-ddl"
version = "1.0.0"
description = "Generate DDL from ERD"
authors = ["Sumit Sahoo"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.0"
gradio = "^4.8.0"
rich = "^13.7.0"
setuptools = "^69.0.2"
openai = "^1.3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
